+++ app/src/apis/teamApi.ts
@@ -15,7 +15,11 @@ interface TeamAPI {
 
 export class FetchTeamAPI implements TeamAPI {
   async getAllTeams(): Promise<Team[]> {
-    const token: any = AuthService.getToken()
+    const token: string | null = AuthService.getToken()
+
+    if (!token) {
+      throw new Error('Token is null')
+    }
     const ownerAddressStore = useOwnerAddressStore()
     const requestOptions = {
       method: 'GET',
@@ -39,7 +43,11 @@ export class FetchTeamAPI implements TeamAPI {
   }
   async getTeam(id: string): Promise<Team | null> {
     const ownerAddressStore = useOwnerAddressStore()
-    const token: any = AuthService.getToken()
+    const token: string | null = AuthService.getToken()
+
+    if (!token) {
+      throw new Error('Token is null')
+    }
     const url = `${BACKEND_URL}/api/teams/${id}`
     const requestOptions = {
       method: 'GET',
@@ -64,7 +72,11 @@ export class FetchTeamAPI implements TeamAPI {
   }
   async updateTeam(id: string, updatedTeamData: Partial<Team>): Promise<Team> {
     const ownerAddressStore = useOwnerAddressStore()
-    const token: any = AuthService.getToken()
+    const token: string | null = AuthService.getToken()
+
+    if (!token) {
+      throw new Error('Token is null')
+    }
 
     const url = `${BACKEND_URL}/api/teams/${id}`
     const requestData = {
@@ -93,7 +105,11 @@ export class FetchTeamAPI implements TeamAPI {
   }
   async deleteTeam(id: string): Promise<void> {
     const url = `${BACKEND_URL}/api/teams/${id}`
-    const token: any = AuthService.getToken()
+    const token: string | null = AuthService.getToken()
+
+    if (!token) {
+      throw new Error('Token is null')
+    }
 
     const requestOptions = {
       method: 'DELETE',
@@ -122,7 +138,11 @@ export class FetchTeamAPI implements TeamAPI {
     teamMembers: Partial<Member>[]
   ): Promise<Team> {
     const ownerAddressStore = useOwnerAddressStore()
-    const token: any = AuthService.getToken()
+    const token: string | null = AuthService.getToken()
+
+    if (!token) {
+      throw new Error('Token is null')
+    }
     const teamObject = {
       name: teamName,
       description: teamDesc,
